Move high-level function to Twitter  API
