+++ src/twitter-api/ObtainingUserDetails.ts
@@ -1,5 +1,5 @@
-import { UserProfileView } from "./twitter-api/ports/UserProfileView";
-import { UserDetails } from "./twitter-model/UserDetails";
+import { UserDetails } from "../twitter-model/UserDetails";
+import { UserProfileView } from "./ports/UserProfileView";
 
 export const obtainingUserDetails = async (
   profileView: UserProfileView,
